Split main.yml into multiple files to support multiple steps on Github Action.
